# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ferrix API (Ajira Copilot)",
    "version": "1.0.0",
    "summary": "Voice-first worker intake, explainable matching, and the human trust pipeline — as a public API.",
    "description": "Ferrix (Ajira Copilot) public API. Auth contract: mutating routes require 'Authorization: Bearer $FERRIX_API_TOKEN' (or 'x-api-key') when that env var is set; unset = open judge demo. EVERY response carries the 'x-ferrix-auth' header (open-demo | enforced). Hand-maintained; kept in lockstep with docs/API.md and the QA suites.",
    "contact": {
      "name": "Ferrix (Ajira Copilot)",
      "url": "https://github.com/"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev server"
    }
  ],
...
  "paths": {
...
    "/api/track-record": {
      "get": {
        "operationId": "getTrackRecord",
        "tags": [
          "Track record"
        ],
        "summary": "Worker's verified work passport (open read)",
        "description": "Verified passport entries (PLACEMENT / TRAINING / REVIEW) for `candidateId`, newest first, plus a summary (counts, average review rating, verified count, last verified date). Unknown candidate → 404, missing query param → 400.",
        "security": [],
        "parameters": [
          {
            "name": "candidateId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "description": "Candidate whose passport to read."
          }
        ],
        "responses": {
          "200": {
            "description": "Entries + summary.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackRecordResponse"
                }
              }
            }
          },
          "400": {
            "description": "`candidateId` query parameter missing.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Candidate not found.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "operationId": "postTrackRecord",
        "tags": [
          "Track record"
        ],
        "summary": "Add a verified passport entry (guarded mutation)",
        "description": "A human coordinator adds a verified entry (placement, completed training, or employer review) to the worker's passport. Entries are stored `verified: true` with `verifiedBy` (defaults to `\"Coordinator (Ujuzi Hub)\"`); `rating` is only stored for REVIEW entries. `occurredAt` accepts any parseable datetime string and defaults to now.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entry created.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "entry"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    },
                    "entry": {
                      "$ref": "#/components/schemas/TrackRecordEntry"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Malformed JSON or body failing the Zod contract (kind enum, title 1–200 chars, rating 1–5, valid `occurredAt`).",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "Candidate not found.",
            "headers": {
              "x-ferrix-auth": {
                "$ref": "#/components/headers/XFerrixAuth"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "FERRIX_API_TOKEN. Required on mutating endpoints only when the deployment sets FERRIX_API_TOKEN (auth mode `enforced`; watch the `x-ferrix-auth` response header). `x-api-key: <token>` is accepted as an equivalent alternative header. In open demo mode (`x-ferrix-auth: open-demo`) no token is needed."
      }
    },
    "headers": {
      "XFerrixAuth": {
        "description": "Auth mode of the deployment for this request: `open-demo` (mutations open) or `enforced` (mutations require the token). Contract: every response carries it — today it is verifiably present on success responses and 401 auth rejections; the shared 400/404/429 helpers are being wired to stamp it as well.",
        "schema": {
          "type": "string",
          "enum": [
            "open-demo",
            "enforced"
          ]
        }
      }
    },
...
    "responses": {
      "Unauthorized": {
        "description": "Auth is enforced and the caller did not present the coordinator token (`Authorization: Bearer <token>` or `x-api-key`).",
        "headers": {
          "x-ferrix-auth": {
            "$ref": "#/components/headers/XFerrixAuth"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected server error.",
        "headers": {
          "x-ferrix-auth": {
            "$ref": "#/components/headers/XFerrixAuth"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Short machine-readable error (first Zod issue message, or a stable label)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Present on 401 auth failures."
          }
        }
      },
...
      "TrackRecordEntry": {
        "type": "object",
        "required": [
          "id",
          "candidateId",
          "kind",
          "title",
          "verified",
          "occurredAt",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "candidateId": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "PLACEMENT",
              "TRAINING",
              "REVIEW"
            ]
          },
          "title": {
            "type": "string"
          },
          "org": {
            "type": [
              "string",
              "null"
            ]
          },
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "rating": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 5,
            "description": "REVIEW entries only."
          },
          "verified": {
            "type": "boolean",
            "const": true
          },
          "verifiedBy": {
            "type": [
              "string",
              "null"
            ]
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TrackRecordResponse": {
        "type": "object",
        "required": [
          "entries",
          "summary"
        ],
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TrackRecordEntry"
            }
          },
          "summary": {
            "type": "object",
            "required": [
              "placements",
              "trainings",
              "reviews",
              "avgRating",
              "verifiedCount",
              "lastVerifiedAt"
            ],
            "properties": {
              "placements": {
                "type": "integer"
              },
              "trainings": {
                "type": "integer"
              },
              "reviews": {
                "type": "integer",
                "description": "REVIEW entries that carry a rating."
              },
              "avgRating": {
                "type": [
                  "number",
                  "null"
                ],
                "description": "Mean review rating, 1 decimal; null when no rated reviews."
              },
              "verifiedCount": {
                "type": "integer"
              },
              "lastVerifiedAt": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "date-time"
              }
            }
          }
        }
      },
      "TrackRecordRequest": {
        "type": "object",
        "required": [
          "candidateId",
          "kind",
          "title"
        ],
        "properties": {
          "candidateId": {
            "type": "string",
            "minLength": 1
          },
          "kind": {
            "type": "string",
            "enum": [
              "PLACEMENT",
              "TRAINING",
              "REVIEW"
            ]
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "org": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 120
          },
          "detail": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 500
          },
          "rating": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 1,
            "maximum": 5
          },
          "occurredAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "verifiedBy": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 120
          }
        }
      },
...
    }
  }
}